requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
streamlit>=1.28.0
PyQt5>=5.15.0
pillow>=10.0.0
